Automatically add published-with-gradle-metadata marker to POM (#26)
If the POM is just the standard project POM, this does not make
it worse. It's even a bit simpler, as the modification is directly
proposed instead of telling the user what to do in an error.

The real advantage of this change is if the POM is generated by another
plugin – in particular the `flatten-maven-plugin`. Then the user does
not have to take care of the marker, which we add automatically to the
POM that the flatten plugin generated for publishing.